Переделал удаление юзера (теперь инфа пишется в отдельную тиблицу), добавил новые Entity для дальнейшнего расширения функционала приложения

+++ find-my-pet/src/main/java/ru/nsu/sberlab/repositories/PetRepository.java
@@ -6,8 +6,8 @@ import ru.nsu.sberlab.models.entities.Pet;
 import java.util.List;
 import java.util.Optional;
 
-public interface PetRepository extends JpaRepository<Pet, Long> {
+public interface PetRepository extends JpaRepository<Pet, Integer> {
     Optional<Pet> findByChipId(String chipId);
     void deleteByChipId(String chipId);
-    List<Pet> findAllByUserId(Long id);
+    List<Pet> findAllByUserId(Integer id);
 }
